norm pass
