+++ include/common.h
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/19 09:43:50 by spyun         #+#    #+#                 */
-/*   Updated: 2025/06/03 15:10:58 by spyun         ########   odam.nl         */
+/*   Updated: 2025/06/05 11:59:21 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -113,9 +113,15 @@ bool			print_error(const char *message, bool return_value);
 bool			print_error_with_value(const char *prefix, const char *value,
 					bool return_value);
 
+/* free_assets.c */
+void			free_asset(t_asset *asset, mlx_t *mlx);
+
 /* free_memory.c */
 void			free_map(t_map *map);
 void			free_game(t_game *game);
+
+/* free_sprites.c */
+void			free_all_sprites(t_game *game);
 
 /* init.c */
 void			init_window(t_game *game);
