+++ src/common/free_memory.c
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/07 11:34:24 by spyun         #+#    #+#                 */
-/*   Updated: 2025/06/05 11:22:48 by spyun         ########   odam.nl         */
+/*   Updated: 2025/06/05 11:58:04 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,79 +27,6 @@ void	free_map(t_map *map)
 	}
 	free(map->grid);
 	map->grid = NULL;
-}
-
-static void	free_asset_paths(t_asset *asset)
-{
-	if (asset->no_path)
-		free(asset->no_path);
-	if (asset->so_path)
-		free(asset->so_path);
-	if (asset->we_path)
-		free(asset->we_path);
-	if (asset->ea_path)
-		free(asset->ea_path);
-	asset->no_path = NULL;
-	asset->so_path = NULL;
-	asset->we_path = NULL;
-	asset->ea_path = NULL;
-}
-
-static void	free_asset_images(t_asset *asset, mlx_t *mlx)
-{
-	if (mlx && asset->no_img)
-		mlx_delete_image(mlx, asset->no_img);
-	if (mlx && asset->so_img)
-		mlx_delete_image(mlx, asset->so_img);
-	if (mlx && asset->we_img)
-		mlx_delete_image(mlx, asset->we_img);
-	if (mlx && asset->ea_img)
-		mlx_delete_image(mlx, asset->ea_img);
-	asset->no_img = NULL;
-	asset->so_img = NULL;
-	asset->we_img = NULL;
-	asset->ea_img = NULL;
-}
-
-static void	free_asset(t_asset *asset, mlx_t *mlx)
-{
-	if (!asset)
-		return ;
-	free_asset_paths(asset);
-	free_asset_images(asset, mlx);
-}
-
-void	free_sprite_frames(mlx_t *mlx, t_sprite *sprite)
-{
-	int	frame;
-	int	max_frames;
-
-	if (sprite->type == SPRITE_TREE)
-		max_frames = TREE_FRAMES;
-	else
-		return ;
-	frame = 0;
-	while (frame < max_frames)
-	{
-		if (sprite->frames[frame])
-		{
-			mlx_delete_image(mlx, sprite->frames[frame]);
-			sprite->frames[frame] = NULL;
-		}
-		frame++;
-	}
-}
-
-void	free_all_sprites(t_game *game)
-{
-	int	i;
-
-	i = 0;
-	while (i < game->map.sprite_count)
-	{
-		free_sprite_frames(game->mlx, &game->map.sprites[i]);
-		i++;
-	}
 }
 
 void	free_game(t_game *game)
